refactor: develop→mainマージ前のコード品質改善 closes #109
- session-lifecycle.ts: セッション作成/クリーンアップの統一（sessions.ts/workspaces.ts/tab.ts→1箇所）
- pane-tree.ts: ペインツリーヘルパー関数の共通化（index.ts/workspaces.ts→1箇所）
- ports.ts: ポート計算定数の一元化
- startup.ts: サーバー起動ロジックの抽出
- worktree-service.ts: execSync→execFileSync（コマンドインジェクション対策）
- tab.ts: claude --continue未実行バグ修正 + execFileSync化
- index.ts: ワークスペース未所属セッションの孤立誤判定バグ修正
- session-store.ts: 不要なeslint-disable除去 + findFirstLeafId重複除去
- workspaces.ts: split-pane失敗時のリソースリーク修正
