+++ packages/server/src/index.ts
@@ -18,6 +18,8 @@ import { createSshRouter } from './routes/ssh.js'
 import { createFeedbackRouter } from './routes/feedback.js'
 import { createWorkspacesRouter } from './routes/workspaces.js'
 import { CanvasStore } from './services/canvas-store.js'
+import { SERVER_PORT_BASE, calculatePort } from './utils/ports.js'
+import { runStartupTasks } from './startup.js'
 
 // PANE_NUMBERからポートを自動算出（develop=0, paneN=N）
 // 設定時は既存PORT環境変数より優先。未設定時のみPORTにフォールバック
@@ -25,7 +27,7 @@ const PANE_NUMBER = process.env.PANE_NUMBER != null
   ? parseInt(process.env.PANE_NUMBER, 10)
   : null
 const PORT = PANE_NUMBER != null
-  ? 14000 + PANE_NUMBER
+  ? calculatePort(SERVER_PORT_BASE, PANE_NUMBER)
   : parseInt(process.env.PORT || '3001', 10)
 const HOST = process.env.HOST || 'localhost'
 
@@ -54,84 +56,8 @@ sshManager.on('exit', (sessionId: string) => {
   markDisconnected(sessionId)
 })
 
-// サーバー起動時: PTYが消失したactiveセッションをdisconnectedに変更
-const orphanedSessions = sessionStore.getAll().filter(s => s.status === 'active')
-if (orphanedSessions.length > 0) {
-  console.log(`⚠️  ${orphanedSessions.length}件のorphanedセッションを検出 → disconnectedに変更`)
-  for (const s of orphanedSessions) {
-    sessionStore.updateStatus(s.id, 'disconnected')
-    console.log(`   ↳ セッション "${s.name}" (${s.id.slice(0, 8)}...) → disconnected`)
-  }
-  console.log('✅ orphanedセッションの復元処理完了。UIから再接続可能です。')
-} else {
-  console.log('✅ orphanedセッションなし')
-}
-
-// サーバー起動時: ペインツリーに含まれない孤立セッションを削除
-try {
-  // 全ワークスペースのペインツリーから参照中のセッションIDを収集
-  const collectSessionIdsFromTree = (node: import('@kurimats/shared').PaneNode): string[] => {
-    if (!node) return []
-    if (node.kind === 'leaf') {
-      return node.surfaces.filter(s => s.type === 'terminal').map(s => s.target)
-    }
-    if (!node.children || node.children.length < 2) return []
-    return [...collectSessionIdsFromTree(node.children[0]), ...collectSessionIdsFromTree(node.children[1])]
-  }
-
-  const workspaces = sessionStore.getAllCmuxWorkspaces()
-  const referencedIds = new Set<string>()
-  for (const ws of workspaces) {
-    for (const id of collectSessionIdsFromTree(ws.paneTree)) {
-      referencedIds.add(id)
-    }
-  }
-
-  // ペインツリーに含まれないセッションを削除（worktreeも含む）
-  const allSessions = sessionStore.getAll()
-  const orphanedCleanup = allSessions.filter(s => !referencedIds.has(s.id))
-  if (orphanedCleanup.length > 0) {
-    console.log(`🧹 ${orphanedCleanup.length}件の孤立セッションを削除します`)
-    for (const s of orphanedCleanup) {
-      if (s.worktreePath && s.repoPath) {
-        try {
-          worktreeService.remove(s.repoPath, s.worktreePath)
-          console.log(`   🗑️ worktree削除: ${s.worktreePath}`)
-        } catch {
-          // 既に削除済みの場合は無視
-        }
-      }
-      sessionStore.delete(s.id)
-      console.log(`   ↳ セッション "${s.name}" (${s.id.slice(0, 8)}...) を削除`)
-    }
-    console.log('✅ 孤立セッション削除完了')
-  } else {
-    console.log('✅ 孤立セッションなし')
-  }
-} catch (e) {
-  console.error('⚠️ 孤立セッション削除中にエラー（サーバー起動は続行）:', e)
-}
-
-// サーバー起動時: worktreeセッションのブランチ名を最新化
-try {
-  const allSessionsForBranch = sessionStore.getAll()
-  let branchFixCount = 0
-  for (const s of allSessionsForBranch) {
-    if (s.worktreePath) {
-      const currentBranch = worktreeService.getBranch(s.worktreePath)
-      if (currentBranch && currentBranch !== s.branch) {
-        sessionStore.updateBranch(s.id, currentBranch)
-        console.log(`   🌿 ブランチ修正: "${s.name}" ${s.branch} → ${currentBranch}`)
-        branchFixCount++
-      }
-    }
-  }
-  if (branchFixCount > 0) {
-    console.log(`✅ ${branchFixCount}件のセッションブランチを修正`)
-  }
-} catch (e) {
-  console.error('⚠️ ブランチ修正中にエラー（サーバー起動は続行）:', e)
-}
+// サーバー起動時の初期化タスク（orphanedセッション復元、孤立削除、ブランチ修正）
+runStartupTasks(sessionStore, worktreeService)
 
 // Express設定
 const app = express()
@@ -156,7 +82,7 @@ app.use('/api/files', createFilesRouter())
 app.use('/api/worktrees', createWorktreesRouter(worktreeService))
 app.use('/api/projects', createProjectsRouter(sessionStore))
 app.use('/api/layout', createLayoutRouter(sessionStore, canvasStore))
-app.use('/api/tab', createTabRouter(sessionStore, ptyManager, sshManager))
+app.use('/api/tab', createTabRouter(sessionStore, ptyManager, sshManager, worktreeService))
 app.use('/api/ssh', createSshRouter(sshManager, sessionStore))
 app.use('/api/feedback', createFeedbackRouter(sessionStore))
 app.use('/api/workspaces', createWorkspacesRouter(sessionStore, ptyManager, sshManager, worktreeService))
